--- custom_components/epa_victoria_air_quality/PyEPA/const.py
@@ -1,15 +1,0 @@
-"""Constants for PyEPA."""
-
-URL_BASE = "https://gateway.api.epa.vic.gov.au/environmentMonitoring/v1/sites/"
-URL_PARAMETERS = "/parameters"
-URL_FIND_SITE = "?environmentalSegment=air&location="
-TIME_SERIES_NAME = "timeSeriesName"
-HOURLY = "1HR_AV"
-DAILY = "24HR_AV"
-READINGS = "readings"
-AVERAGE_VALUE = "averageValue"
-HEALTH_ADVICE = "healthAdvice"
-TIME_SERIES_READINGS = "timeSeriesReadings"
-UNTIL = "until"
-PARAMETERS = "parameters"
-CONFIDENCE = "confidence"
