+++ custom_components/epa_victoria_air_quality/const.py
@@ -13,7 +13,6 @@ CONF_SITE_ID = "site_id"
 DOMAIN = "epa_victoria_air_quality"
 ENTRY_TYPE_SERVICE: Final = "service"
 COORDINATOR: Final = "coordinator"
-EPA_URL = "https://gateway.api.epa.vic.gov.au/environmentMonitoring/v1/sites/"
 TIME_FORMAT = "%H:%M:%S"
 DATE_FORMAT = "%Y-%m-%d %H:%M:%S"
 DATE_FORMAT_UTC = "%Y-%m-%d %H:%M:%S UTC"
@@ -31,3 +30,16 @@ TYPE_AQI_PM25 = "aqi_pm25"
 TYPE_AQI_PM25_24H = "aqi_pm25_24h"
 TYPE_PM25 = "pm25"
 TYPE_PM25_24H = "pm25_24h"
+URL_BASE = "https://gateway.api.epa.vic.gov.au/environmentMonitoring/v1/sites/"
+URL_PARAMETERS = "/parameters"
+URL_FIND_SITE = "?environmentalSegment=air&location="
+TIME_SERIES_NAME = "timeSeriesName"
+HOURLY = "1HR_AV"
+DAILY = "24HR_AV"
+READINGS = "readings"
+AVERAGE_VALUE = "averageValue"
+HEALTH_ADVICE = "healthAdvice"
+TIME_SERIES_READINGS = "timeSeriesReadings"
+UNTIL = "until"
+PARAMETERS = "parameters"
+CONFIDENCE = "confidence"
